+++ cores/epoxy/Arduino.h
@@ -16,8 +16,8 @@
 #include <epoxy_test/ArduinoTest.h>
 
 // xx.yy.zz => xxyyzz (without leading 0)
-#define EPOXY_DUINO_VERSION 10500
-#define EPOXY_DUINO_VERSION_STRING "1.5.0"
+#define EPOXY_DUINO_VERSION 10600
+#define EPOXY_DUINO_VERSION_STRING "1.6.0"
 
 #include <algorithm> // min(), max()
 #include <cmath> // abs()