+++ cores/epoxy/Esp.h
@@ -25,6 +25,8 @@ class EspClass
 
     uint32_t getCpuFreqMHZ() { return 80; }
 
+    uint32_t getChipId() { return 0; }
+
     uint32_t getCycleCount() {
       struct timeval now;
       gettimeofday(&now, NULL);